requirements.txt
streamlit
pypandoc-binary
fpdf2
beautifulsoup4